daemon: move sync `jsonrpc_` methods to an individual module
