+++ lbry/extras/daemon/daemon.py
@@ -234,6 +234,7 @@ from lbry.extras.daemon.daemon_get import Daemon_get
 from lbry.extras.daemon.daemon_settings import Daemon_settings
 from lbry.extras.daemon.daemon_wallet import Daemon_wallet
 from lbry.extras.daemon.daemon_account import Daemon_account
+from lbry.extras.daemon.daemon_sync import Daemon_sync
 
 HISTOGRAM_BUCKETS = (
     .005, .01, .025, .05, .075, .1, .25, .5, .75, 1.0, 2.5, 5.0, 7.5, 10.0, 15.0, 20.0, 30.0, 60.0, float('inf')
@@ -241,7 +242,7 @@ HISTOGRAM_BUCKETS = (
 
 
 class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
-             Daemon_account):
+             Daemon_account, Daemon_sync):
     """
     LBRYnet daemon, a jsonrpc interface to lbry functions
     """
@@ -793,71 +794,8 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
     SYNC_DOC = """
     Wallet synchronization.
     """
-
-    @requires("wallet")
-    def jsonrpc_sync_hash(self, wallet_id=None):
-        """
-        Deterministic hash of the wallet.
-
-        Usage:
-            sync_hash [<wallet_id> | --wallet_id=<wallet_id>]
-
-        Options:
-            --wallet_id=<wallet_id>   : (str) wallet for which to generate hash
-
-        Returns:
-            (str) sha256 hash of wallet
-        """
-        wallet = self.wallet_manager.get_wallet_or_default(wallet_id)
-        return hexlify(wallet.hash).decode()
-
-    @requires("wallet")
-    async def jsonrpc_sync_apply(self, password, data=None, wallet_id=None, blocking=False):
-        """
-        Apply incoming synchronization data, if provided, and return a sync hash and update wallet data.
-
-        Wallet must be unlocked to perform this operation.
-
-        If "encrypt-on-disk" preference is True and supplied password is different from local password,
-        or there is no local password (because local wallet was not encrypted), then the supplied password
-        will be used for local encryption (overwriting previous local encryption password).
-
-        Usage:
-            sync_apply <password> [--data=<data>] [--wallet_id=<wallet_id>] [--blocking]
-
-        Options:
-            --password=<password>         : (str) password to decrypt incoming and encrypt outgoing data
-            --data=<data>                 : (str) incoming sync data, if any
-            --wallet_id=<wallet_id>       : (str) wallet being sync'ed
-            --blocking                    : (bool) wait until any new accounts have sync'ed
-
-        Returns:
-            (map) sync hash and data
-
-        """
-        wallet = self.wallet_manager.get_wallet_or_default(wallet_id)
-        wallet_changed = False
-        if data is not None:
-            added_accounts = wallet.merge(self.wallet_manager, password, data)
-            if added_accounts and self.ledger.network.is_connected:
-                if blocking:
-                    await asyncio.wait([
-                        a.ledger.subscribe_account(a) for a in added_accounts
-                    ])
-                else:
-                    for new_account in added_accounts:
-                        asyncio.create_task(self.ledger.subscribe_account(new_account))
-            wallet_changed = True
-        if wallet.preferences.get(ENCRYPT_ON_DISK, False) and password != wallet.encryption_password:
-            wallet.encryption_password = password
-            wallet_changed = True
-        if wallet_changed:
-            wallet.save()
-        encrypted = wallet.pack(password)
-        return {
-            'hash': self.jsonrpc_sync_hash(wallet_id),
-            'data': encrypted.decode()
-        }
+    # jsonrpc_sync_hash
+    # jsonrpc_sync_apply
 
     ADDRESS_DOC = """
     List, generate and verify addresses.
